requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.9.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
distro==1.9.0
Flask==3.1.0
Flask-Cors==5.0.0
groq==0.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
langdetect==1.0.9
MarkupSafe==3.0.2
pydantic==2.10.2
pydantic_core==2.27.1
PyJWT==2.10.1
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
typing_extensions==4.12.2
Werkzeug==3.1.3
